refactor: extract flattenSchema to lib/schema.ts
Separate browser-safe schema utils from fs-dependent openapi parser.
